enp/requirements.txt
six
django
django-crispy-forms
django-sslserver
django-shortuuidfield